Export dacapo

+++ dcc/model_export/generate_metadata.py
@@ -3,9 +3,13 @@ import click
 from typing import List, Optional
 from pydantic import BaseModel, Field
 import os
+import dcc.model_export.config as c
 
 
 def get_export_folder():
+    if c.DCC_EXPORT_FOLDER is not None:
+        return c.DCC_EXPORT_FOLDER
+    print("DCC_EXPORT_FOLDER is not set in the config, checking environment variables...")
     folder = os.getenv("DCC_EXPORT_FOLDER")
     if not folder:
         folder = input(
@@ -16,29 +20,29 @@ def get_export_folder():
 
 
 class ModelMetadata(BaseModel):
-    model_name: Optional[str] = None
-    model_type: Optional[str] = Field(None, description="UNet or DenseNet121")
-    framework: Optional[str] = Field(None, description="MONAI or PyTorch")
-    spatial_dims: Optional[int] = Field(None, description="2 or 3")
-    in_channels: Optional[int] = None
-    out_channels: Optional[int] = None
-    channels_names: Optional[List[str]] = Field(
-        None, description="Comma-separated values, e.g., 'CT, PET'"
-    )
-    input_shape: Optional[List[int]] = Field(
-        None, description="Comma-separated values, e.g., 1,1,96,96,96"
-    )
-    output_shape: Optional[List[int]] = Field(
-        None, description="Comma-separated values, e.g., 1,2,96,96,96"
-    )
-    author: Optional[str] = None
-    description: Optional[str] = None
-    version: Optional[str] = "1.0.0"
+    model_name: Optional[str] = Field(None, description="Name of the model")
+    model_type: Optional[str] = Field(None, description="Type of the model, e.g., UNet or DenseNet121")
+    framework: Optional[str] = Field(None, description="Framework used, e.g., MONAI or PyTorch")
+    spatial_dims: Optional[int] = Field(None, description="Number of spatial dimensions, e.g., 2 or 3")
+    in_channels: Optional[int] = Field(None, description="Number of input channels")
+    out_channels: Optional[int] = Field(None, description="Number of output channels")
+    iteration: Optional[int] = Field(None, description="Iteration number")
+    input_voxel_size: Optional[List[int]] = Field(None, description="Input voxel size as comma-separated values, e.g., 8,8,8")
+    output_voxel_size: Optional[List[int]] = Field(None, description="Output voxel size as comma-separated values, e.g., 8,8,8")
+    channels_names: Optional[List[str]] = Field(None, description="Names of the channels as comma-separated values, e.g., 'CT, PET'")
+    input_shape: Optional[List[int]] = Field(None, description="Input shape as comma-separated values, e.g., 1,1,96,96,96")
+    output_shape: Optional[List[int]] = Field(None, description="Output shape as comma-separated values, e.g., 1,2,96,96,96")
+    inference_input_shape: Optional[List[int]] = Field(None, description="Inference input shape as comma-separated values, e.g., 1,1,96,96,96")
+    inference_output_shape: Optional[List[int]] = Field(None, description="Inference output shape as comma-separated values, e.g., 1,2,96,96,96")
+    author: Optional[str] = Field(None, description="Author of the model")
+    description: Optional[str] = Field(None, description="Description of the model")
+    version: Optional[str] = Field("1.0.0", description="Version of the model")
 
 
 def generate_readme(metadata: ModelMetadata):
     readme_content = f"""
     # {metadata.model_name} Model
+    iteration: {metadata.iteration}
 
     ## Description
     {metadata.description}
@@ -62,12 +66,19 @@ def generate_readme(metadata: ModelMetadata):
     return readme_content
 
 
-def export_metadata(metadata: ModelMetadata):
-    prompt_for_missing_fields(metadata)
+def export_metadata(metadata: ModelMetadata, overwrite: bool = False):
+    
     export_folder = get_export_folder()
     result_folder = os.path.join(export_folder, metadata.model_name)
+    if os.path.exists(result_folder) and not overwrite:
+        result = click.confirm(
+            f"Folder {result_folder} already exists. Do you want to overwrite it?",
+        )
+        if not result:
+            return
+    metadata = prompt_for_missing_fields(metadata)
     os.makedirs(result_folder, exist_ok=True)
-    output_file = os.path.join(result_folder, f"{metadata.model_name}_metadata.json")
+    output_file = os.path.join(result_folder, "metadata.json")
     with open(output_file, "w") as f:
         json.dump(metadata.dict(), f, indent=4)
     click.echo(f"Metadata saved to {output_file}")
@@ -82,20 +93,26 @@ def prompt_for_missing_fields(metadata: ModelMetadata):
     for field_name, field in metadata.__fields__.items():
         value = getattr(metadata, field_name)
         if value is None:
-            prompt_text = (
-                field.field_info.description or f"Enter {field_name.replace('_', ' ')}"
-            )
-            if field_name in ["channels_names", "input_shape", "output_shape"]:
-                user_input = click.prompt(prompt_text, type=str)
-                if field_name == "channels_names":
-                    value = [item.strip() for item in user_input.split(",")]
+            try:
+                prompt_text = (
+                field.description or f"Enter {field_name.replace('_', ' ')}"
+                )
+
+                if field_name in ["channels_names", "input_shape", "output_shape"]:
+                    user_input = click.prompt(prompt_text, type=str)
+                    if field_name == "channels_names":
+                        value = [item.strip() for item in user_input.split(",")]
+                    else:
+                        value = [int(item) for item in user_input.split(",")]
+                elif field.type_ == int:
+                    value = click.prompt(prompt_text, type=int)
                 else:
-                    value = [int(item) for item in user_input.split(",")]
-            elif field.type_ == int:
-                value = click.prompt(prompt_text, type=int)
-            else:
-                value = click.prompt(prompt_text, type=str)
-            setattr(metadata, field_name, value)
+                    value = click.prompt(prompt_text, type=str)
+                setattr(metadata, field_name, value)
+            except Exception as e:
+                raise Exception(f"Field {field_name} is missing a description. {e}")
+
+    return metadata
 
 
 @click.command()
